creating robust example of the api
